requirements.txt
tensorflow==2.4.1
matplotlib==3.3.4
numpy==1.19.3
Pillow==8.2.0
